+++ amazon-chime-sdk/src/main/java/com/amazonaws/services/chime/sdk/meetings/audiovideo/AudioVideoConfiguration.kt
@@ -6,7 +6,7 @@
 package com.amazonaws.services.chime.sdk.meetings.audiovideo
 
 import com.amazonaws.services.chime.sdk.meetings.audiovideo.audio.AudioMode
-import com.amazonaws.services.chime.sdk.meetings.audiovideo.audio.AudioStream
+import com.amazonaws.services.chime.sdk.meetings.audiovideo.audio.AudioStreamType
 
 /**
  * [AudioVideoConfiguration] represents the configuration to be used for audio and video during a
@@ -15,10 +15,10 @@ import com.amazonaws.services.chime.sdk.meetings.audiovideo.audio.AudioStream
  * @property audioMode: AudioMode - the audio mode in which the audio client should operate during
  * a meeting session.
  *
- * @property audioStream: AudioStream - the audio stream in which the audio client should operate
- * during a meeting session.
+ * @property audioStreamType: AudioStreamType - the audio stream type in which the audio client
+ * should operate during a meeting session.
  */
 data class AudioVideoConfiguration @JvmOverloads constructor(
     val audioMode: AudioMode = AudioMode.Stereo48K,
-    val audioStream: AudioStream = AudioStream.VoiceCall
+    val audioStreamType: AudioStreamType = AudioStreamType.VoiceCall
 )